Update the API bindings
