+++ XenModel/XenAPI/Helper.cs
@@ -35,128 +35,9 @@ using System.Collections.Generic;
 
 namespace XenAPI
 {
-    public enum API_Version
-    {
-        API_1_1 = 1,  // XenServer 4.0 (codename Rio)
-        API_1_2 = 2,  // XenServer 4.1 (Miami)
-        API_1_3 = 3,  // XenServer 5.0 up to update 2 (Orlando)
-        API_1_4 = 4,  // Unreleased
-        API_1_5 = 5,  // XenServer 5.0 update 3 and above (Floodgate)
-        API_1_6 = 6,  // XenServer 5.5 (George)
-        API_1_7 = 7,  // XenServer 5.6 (Midnight Ride)
-        API_1_8 = 8,  // XenServer 5.6.1 (Cowley)
-        API_1_9 = 9,  // XenServer 6.0 (Boston)
-        API_1_10 = 10, // XenServer 6.1 (Tampa)
-        API_2_0 = 11, // XenServer 6.2 (Clearwater)
-        API_2_1 = 12, // XenServer 6.2 with vGPU (vGPU)
-        API_2_2 = 13, // XenServer 6.2 Hotfix XS62ESP1004 (Felton)
-        API_2_3 = 14, // XenServer Creedence
-        API_2_4 = 15, // XenServer Cream
-        API_2_5 = 16, // XenServer Dundee
-        API_2_6 = 17, // XenServer Ely
-        API_2_7 = 18, // XenServer Falcon
-        LATEST = 18,
-        // Don't forget to change LATEST above, and APIVersionString below.
-        UNKNOWN = 99
-    }
-
-    public static class Helper
+    public static partial class Helper
     {
         public const string NullOpaqueRef = "OpaqueRef:NULL";
-
-        public static string APIVersionString(API_Version v)
-        {
-            switch (v)
-            {
-                case API_Version.API_1_1:
-                    return "1.1";
-                case API_Version.API_1_2:
-                    return "1.2";
-                case API_Version.API_1_3:
-                    return "1.3";
-                case API_Version.API_1_4:
-                    return "1.4";
-                case API_Version.API_1_5:
-                    return "1.5";
-                case API_Version.API_1_6:
-                    return "1.6";
-                case API_Version.API_1_7:
-                    return "1.7";
-                case API_Version.API_1_8:
-                    return "1.8";
-                case API_Version.API_1_9:
-                    return "1.9";
-                case API_Version.API_1_10:
-                    return "1.10";
-                case API_Version.API_2_0:
-                    return "2.0";
-                case API_Version.API_2_1:
-                    return "2.1";
-                case API_Version.API_2_2:
-                    return "2.2";
-                case API_Version.API_2_3:
-                    return "2.3";
-                case API_Version.API_2_4:
-                    return "2.4";
-                case API_Version.API_2_5:
-                    return "2.5";
-                case API_Version.API_2_6:
-                    return "2.6";
-                case API_Version.API_2_7:
-                    return "2.7";
-                default:
-                    return "Unknown";
-            }
-        }
-
-        public static API_Version GetAPIVersion(long major, long minor)
-        {
-            try
-            {
-                return (API_Version)Enum.Parse(typeof(API_Version),
-                    string.Format("API_{0}_{1}", major, minor));
-            }
-            catch (ArgumentException)
-            {
-                return API_Version.UNKNOWN;
-            }
-        }
-
-        /// <summary>
-        /// Converts the string representation of an API version number to its API_Version equivalent.
-        /// This function assumes that API version numbers are of form a.b
-        /// </summary>
-        public static API_Version GetAPIVersion(string version)
-        {
-            if (version != null)
-            {
-                string[] tokens = version.Split('.');
-                int major, minor;
-                if (tokens.Length == 2 && int.TryParse(tokens[0], out major) && int.TryParse(tokens[1], out minor))
-                {
-                    return GetAPIVersion(major, minor);
-                }
-            }
-            return API_Version.UNKNOWN;
-        }
-
-        /// <summary>
-        /// Return a positive number if the given session's API version is greater than the given
-        /// API_version, negative if it is less, and 0 if they are equal.
-        /// </summary>
-        internal static int APIVersionCompare(Session session, API_Version v)
-        {
-            return (int)session.APIVersion - (int)v;
-        }
-
-        /// <summary>
-        /// Return true if the given session's API version is greater than or equal to the given
-        /// API_version.
-        /// </summary>
-        internal static bool APIVersionMeets(Session session, API_Version v)
-        {
-            return APIVersionCompare(session, v) >= 0;
-        }
 
         /// <summary>
         /// Test to see if two objects are equal.  If the objects implement ICollection, then we will
